fix: use absolute path in records
The problem happens on citre-get-field. Ask it for "path", it will call
citre--project-root. And when asking for "line", it asks itself for
"path" in a temp buffer, so citre--project-root is called in that temp
buffer, which is somewhere out of the project. We use absolute path
directly in the records, so upper components don't need to care about
this.

The "file" field in records is removed. The thing is it's not guaranteed
that we can get relative paths in ctags output. From my understanding,
it outputs relative path only when using the -R option. A function
citre--relative-path is added to get path relative to the project root
afterwards.