refactor: in basic functions/APIs, use root dir to specify project
The original way is to specify it by a buffer, then the project of that
buffer is used. But it's rare that we need to explicitely specify the
project root, and when this happens, we should have more direct control
to suit different needs. The docstrings and messages are also updated to
reflect this change.