refactor: use readtags for citre--get-lines
This is an important step. Through readtags, we can make use of fields
like scope and inherits, then it's possible to offer semantic completion
(like completing struct member or instance attribute/method). Another
plan is to offer smart sorting for jump/completion candidates. We guess
which are probably wanted by the user, and put them on the top.

Currently we still get the lines only by matching the names of tags. The
feature of using Citre without a tags file is removed, since to
implement the above features, we will need to run readtags multiple
times, then it's better to use a temporary file instead. This will be
done later.