refactor: get rid of the project concept in core layer
The core layer should focus on interact with readtags output. We rule
out the project concept so the functions are purer, means more reliable
and easier to test.

We still need to remove the PROJECT argument in citre-get-records. We
could use the _TAG_PROC_CWD pseudo tag to achieve what we want.

This is an incompatible change for upper tools. The functions whose
signatures change, and we should pay attention to, are:
citre-get-records, citre-get-definition-locations, and
citre-get-completions.

See #10.