tweak: use hash tables for additional info
The original implementation uses a plain list, and a lookup table to
translate the field to the index. In average, this is as fast as a hash
table.

But hash table makes it easier to work with place forms: You can't
automatically grow a list when setf a Nth element which is not in the
list, so we need to hard code the length of the list in some initialize
code. This is not needed for alists or hash tables, and we need hash
tables due to performance reasons.