enh: offer ext-lang field
This extension field guesses the language based on the file extension
when the language field is missing. This is a byproduct of implementing
the kind-full extension field.

A change in the line parsing is introduced for this: all the fields that
extension fields depend on is made optional. Since the dependency
relationship can actually be complex (like in ext-lang field, it
requires the language field OR the input field), and whether the needed
fields are missing can be checked when actually calculating the
extension field, so it's simpler to make them optional and don't throw
an error when parsing fields in the line.