core, perf: get additional tags file info in one pass
The original idea is that different tools needs different extension
fields, which requires different pieces of info, so we only fetch those
info on demand. But since there are not many kinds of additional info,
the benefit is not obvious, so we choose to get them in one pass, to
reduce shell command calls, and shorten the "cold boot" time when we use
a tags file for the first time.