tool, feat: handle code reading history in citre-peek (#35)
citre-peek-through command is added. It lets you pick a symbol in the
peek window in the ace-jump fashion, and peek the definition of that
symbol.

This means we have a history in a peek session. Data structures are
created to handle the tree-like history, and some commands are added to
browse/modify the tree.

Now we can go down the rabbit hole without even leaving current buffer!