core, refactor: switch to another method to poll for async process
The old method uses accept-process-output. It seems to be the canonical
way, but in fact, we can't perform local quit in it, see
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=32986.

Thus, while-no-input can't escape from it. If it does, it's because
inteference from timers. Surprisingly this "accidental" mechamism works
perfectly on GNU/Linux, but recent bug reports (#61, #69, #70) show the
problem on Windows.

The new scheme uses a combination of catch/throw, and sit-for/sleep-for
tricks. This should guarantee that local quits (by user input or C-g)
will success whenever Emacs allows it.