core, fix: don't modify string arguments come from upper components
If a string argument is passed by upper components (i.e., not from the
tags file), there are chances that it references variables that may be
used later, so we shouldn't modify it.

A bug related to it is on Windows, when generating filter for input
field, the disk symbol of (buffer-file-name) is upcased.