test, fix: set file name for get-definitions; fix C tests
When a tag has file scope, and is not in current file, Citre filters
them out from definitions.

For this to work in the tests, we should set buffer-file-name after load
a file into the buffer.

This causes problems in existing C tests. The solution is change the
extension from .c to .h so the file-local definitions become public.