common, add: async process infrastructure
Async process infrastructure is added so we have a basis for
interactively inspecting a tags file, which is a planned future feature.

The synchronous process interface is also rewritten based on this. It's
cleaner than the previous implementation, and hopefully could solve
process related problems on Windows, which I constantly get complaints
about from other forums.