refactor: rename citre-core library to citre-readtags
This is the first step towards an overall refactoring. The goal is to:

1. Separate the modules clearly, and reducing the dependencies between them.

2. Create libraries for the UI, including citre-jump, citre-peek, and
   capf/xref/imenu integration. They should depend on the "tag" data structure
   but not the readtags program.

3. On the basis of 2, build a pluggable backend structure, so that readtags is
   one of the backends. This could recude the amount of code supporting
   readtags and make it easier to maintain.

Though we may introduce other backends (as long as it's easy to implement and
maintain), the readtags backend will still be the center of development.

In this commit, we'd like to make it clear that citre-core is a readtags
library, and it may be inadequate to call it the core.