tags, fix: use a separate file to record files to scan
In commit 615da54, we switched to using an option file to record command
line options, and I ignored the fact that an option file can't record
the files to scan. The result is: only the project root is scanned.

Now we add a separated file to record files to scan, and the "tags file
generating recipe" editing tool is rewritten to make this easy to use.